+++ lib/src/vroute_elements/vnester.dart
@@ -137,6 +137,14 @@ class VNester extends VRouteElementBuilder {
   final Widget Function(Animation<double> animation,
       Animation<double> secondaryAnimation, Widget child)? buildTransition;
 
+  /// A key for the nested navigator
+  /// It is created automatically
+  ///
+  /// Using this is useful if you create two different [VNesterPageBase] that should
+  /// actually be the same. This happens if you use two different [VRouteElementBuilder]
+  /// to represent two different routes which should share a common [VNesterPageBase]
+  final GlobalKey<NavigatorState>? navigatorKey;
+
   VNester({
     required this.path,
     required this.widgetBuilder,
@@ -149,6 +157,7 @@ class VNester extends VRouteElementBuilder {
     this.stackedRoutes = const [],
     this.aliases = const [],
     this.mustMatchStackedRoute = false,
+    this.navigatorKey,
   });
 
   @override
@@ -167,6 +176,7 @@ class VNester extends VRouteElementBuilder {
               buildTransition: buildTransition,
               transitionDuration: transitionDuration,
               reverseTransitionDuration: reverseTransitionDuration,
+              navigatorKey: navigatorKey,
             ),
           ],
         ),